requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.4
kiwisolver==1.4.8
matplotlib==3.10.3
mysql-connector-python==9.3.0
numpy==2.3.1
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
python-dateutil==2.9.0.post0
six==1.17.0
